Cleaned up imports and renamed some js file to jsx because they use JSX syntax.
