+++ meteor-app/imports/ui/components/afInputLocation/afInputLocation.js
@@ -1,4 +1,3 @@
-import { Meteor } from "meteor/meteor";
 import { Template } from "meteor/templating";
 import { AutoForm } from "meteor/aldeed:autoform";
