+++ meteor-app/imports/ui/components/companyRatingsComponent.jsx
@@ -3,15 +3,12 @@ import StarRatings from "react-star-ratings";
 import PropTypes from "prop-types";
 import { withStyles } from "@material-ui/core/styles";
 import Card from "@material-ui/core/Card";
-import CardActions from "@material-ui/core/CardActions";
-import CardContent from "@material-ui/core/CardContent";
-import CardActionArea from "@material-ui/core/CardActionArea";
 import CardMedia from "@material-ui/core/CardMedia";
-import Button from "@material-ui/core/Button";
 import Typography from "@material-ui/core/Typography";
 import CircularProgressbar from "react-circular-progressbar";
-import i18n from "meteor/universe:i18n";
 import { createMuiTheme } from "@material-ui/core";
+
+import i18n from "meteor/universe:i18n";
 
 const T = i18n.createComponent();
