+++ meteor-app/imports/ui/components/showJobComponent.jsx
@@ -9,7 +9,7 @@ import { Link } from "react-router-dom";
 
 import i18n from "meteor/universe:i18n";
 
-import { processLocation } from "/imports/api/models/helpers/postgresql/misc.js";
+import { processLocation } from "/imports/api/models/helpers/postgresql/misc.ts";
 import { urlGenerators } from "/imports/ui/pages";
 import withUpdateOnChangeLocale from "/imports/ui/hoc/update-on-change-locale.jsx";
