+++ meteor-app/imports/ui/pages/apply-for-job/apply-for-job.jsx
@@ -1,19 +1,17 @@
-// Boilerplate first
-import { Meteor } from "meteor/meteor";
 import React from "react";
 import PropTypes from "prop-types";
-import { Template } from "meteor/templating"; // Used to set up the autoform
-import Blaze from "meteor/gadicc:blaze-react-component"; // used to insert Blaze templates into React components
-import ErrorWidget from "/imports/ui/components/error-widget.jsx"; // used to display errors thrown by methods
-import { ReactiveVar } from "meteor/reactive-var";
-import { ReactiveDict } from "meteor/reactive-dict"; // used to hold global state because...you can't "pass props" to Blaze templates
+
+import { Meteor } from "meteor/meteor";
+import { Template } from "meteor/templating";
+import Blaze from "meteor/gadicc:blaze-react-component";
+import { ReactiveDict } from "meteor/reactive-dict";
 import { AutoForm } from "meteor/aldeed:autoform";
 import i18n from "meteor/universe:i18n";
 
+import ErrorWidget from "/imports/ui/components/error-widget.jsx";
 import PageWrapper from "/imports/ui/components/page-wrapper";
-
-// Specific stuff second
 import { JobApplicationSchema } from "/imports/api/data/jobads.js";
+
 import "./apply-for-job.html";
 
 const afj_form_state = new ReactiveDict();
