+++ meteor-app/imports/ui/pages/company-profile/tabs/jobs/job-posting.jsx
@@ -10,7 +10,7 @@ import { Link } from "react-router-dom";
 import i18n from "meteor/universe:i18n";
 
 import { urlGenerators } from "/imports/ui/pages";
-import { processLocation } from "/imports/api/models/helpers/postgresql/misc.js";
+import { processLocation } from "/imports/api/models/helpers/postgresql/misc.ts";
 import withUpdateOnChangeLocale from "/imports/ui/hoc/update-on-change-locale.jsx";
 
 const T = i18n.createComponent();
