+++ meteor-app/imports/ui/pages/company-profile/tabs/salaries/salaries.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPlus } from "@fortawesome/free-solid-svg-icons";
-import { Link } from "react-router-dom";
 
 import i18n from "meteor/universe:i18n";
