+++ meteor-app/imports/ui/pages/company-profile/tabs/sections/job-ads.jsx
@@ -9,7 +9,7 @@ import { Link } from "react-router-dom";
 
 import i18n from "meteor/universe:i18n";
 
-import { processLocation } from "/imports/api/models/helpers/postgresql/misc.js";
+import { processLocation } from "/imports/api/models/helpers/postgresql/misc.ts";
 import { urlGenerators } from "/imports/ui/pages";
 
 const T = i18n.createComponent();
