+++ meteor-app/imports/ui/pages/company-search/company-search.jsx
@@ -1,16 +1,17 @@
-import { Meteor } from "meteor/meteor";
 import React from "react";
 import PropTypes from "prop-types";
-import i18n from "meteor/universe:i18n";
 import { Query } from "react-apollo";
-import { processLocation } from "/imports/api/models/helpers/postgresql/misc.js";
+
+import i18n from "meteor/universe:i18n";
+
 import PageWrapper from "/imports/ui/components/page-wrapper";
 import CompanySearchResult from "/imports/ui/components/company-search-result.jsx";
 import CompaniesSearchBar from "/imports/ui/components/companies-search-bar.jsx";
-import withUpdateOnChangeLocale from "/imports/ui/hoc/update-on-change-locale.jsx";
-import companySearchQuery from "./company-search.graphql";
-import Spinner from "../../components/Spinner";
+import Spinner from "/imports/ui/components/Spinner";
 import PaginateSystem from "/imports/ui/components/paginate/pagination.jsx";
+import withUpdateOnChangeLocale from "/imports/ui/hoc/update-on-change-locale.jsx";
+
+import companySearchQuery from "./company-search.graphql";
 
 const t = i18n.createTranslator("common.search");
 const T = i18n.createComponent(t);
