+++ meteor-app/imports/ui/pages/post-a-job/post-a-job.jsx
@@ -1,20 +1,19 @@
-// Boilerplate first
-import { Meteor } from "meteor/meteor";
 import React from "react";
-import { Template } from "meteor/templating"; // Used to set up the autoform
-import Blaze from "meteor/gadicc:blaze-react-component"; // used to insert Blaze templates into React components
-import ErrorWidget from "/imports/ui/components/error-widget.jsx"; // used to display errors thrown by methods
-import { ReactiveDict } from "meteor/reactive-dict"; // used to hold global state because...you can't "pass props" to Blaze templates
+
+import { Meteor } from "meteor/meteor";
+import { Template } from "meteor/templating";
+import Blaze from "meteor/gadicc:blaze-react-component";
+import { ReactiveDict } from "meteor/reactive-dict";
 import { AutoForm } from "meteor/aldeed:autoform";
 import i18n from "meteor/universe:i18n";
 
-// Specific stuff second
+import { reactiveCommonTranslator } from "/imports/ui/startup/i18n.js";
 import { JobAdSchema } from "/imports/api/data/jobads.js";
-import "./post-a-job.html";
-
+import ErrorWidget from "/imports/ui/components/error-widget.jsx";
 import PageWrapper from "/imports/ui/components/page-wrapper";
-
 import "/imports/ui/components/afInputLocation";
+
+import "./post-a-job.html";
 
 const paj_form_state = new ReactiveDict();
 paj_form_state.set("formError", {
@@ -56,8 +55,6 @@ paj_form_state.set("contractTypeOptions", [
 ]);
 
 if (Meteor.isClient) {
-	import { reactiveCommonTranslator } from "/imports/ui/startup/i18n.js";
-
 	Template.paj_blaze_form.bindI18nNamespace("common.forms");
 
 	Template.paj_blaze_form.onCreated(function() {
