+++ meteor-app/imports/ui/pages/register/register-form.jsx
