+++ meteor-app/server/main.js
@@ -1,16 +1,14 @@
 import { Meteor } from "meteor/meteor";
 import { WebApp } from "meteor/webapp";
 
-import { app, onServerReady } from "/imports/api/http";
+import { app, onServerReady } from "/imports/api/http/index.ts";
 // Import all of the collection and method definitions.
 // These files need to be run to initialize the MongoDB connections.
 import "/imports/api/data";
-import { sendEmail } from "/imports/api/connectors/email.ts";
 
 // install the express server within meteor webapp connect
 WebApp.rawConnectHandlers.use(app);
 
 Meteor.startup(() => {
 	onServerReady();
-	//sendEmail("checking");
 });
